backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
boto3==1.34.34
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
